+++ src/main/java/org/bytedeco/javacpp/Pointer.java
@@ -294,6 +294,11 @@ public class Pointer implements AutoCloseable {
      * Set via "org.bytedeco.javacpp.maxbytes" system property, defaults to {@link Runtime#maxMemory()}. */
     static final long maxBytes;
 
+    /** Maximum number of times to call {@link System#gc()} before giving up with {@link OutOfMemoryError}.
+     * Set via "org.bytedeco.javacpp.maxretries" system property, defaults to 10, where each retry is followed
+     * by a call to {@code Thread.sleep(100)}. */
+    static final int maxRetries;
+
     static {
         String s = System.getProperty("org.bytedeco.javacpp.nopointergc", "false").toLowerCase();
         if (s.equals("true") || s.equals("t") || s.equals("")) {
@@ -331,6 +336,17 @@ public class Pointer implements AutoCloseable {
             }
         }
         maxBytes = m;
+
+        int n = 10;
+        s = System.getProperty("org.bytedeco.javacpp.maxretries");
+        if (s != null && s.length() > 0) {
+            try {
+                n = Integer.parseInt(s);
+            } catch (NumberFormatException e) {
+                // keep default value set above
+            }
+        }
+        maxRetries = n;
     }
 
     /** Clears, deallocates, and removes all garbage collected objects from the {@link #referenceQueue}. */
@@ -452,7 +468,7 @@ public class Pointer implements AutoCloseable {
                     (DeallocatorReference)deallocator :
                     new DeallocatorReference(this, deallocator);
             int count = 0;
-            while (count++ < 10 && maxBytes > 0 && DeallocatorReference.totalBytes + r.bytes > maxBytes) {
+            while (count++ < maxRetries && maxBytes > 0 && DeallocatorReference.totalBytes + r.bytes > maxBytes) {
                 try {
                     // try to get some more memory back
                     System.gc();
@@ -631,7 +647,8 @@ public class Pointer implements AutoCloseable {
     /**
      * Checks for equality with argument. Defines obj to be equal if {@code
      *     (obj == null && this.address == 0) ||
-     *     (obj.address == this.address && obj.position == this.position)}.
+     *     (obj.address == this.address && obj.position == this.position)},
+     * and the classes are the same, unless one of them in Pointer itself.
      *
      * @param obj the object to compare this Pointer to
      * @return true if obj is equal
@@ -641,7 +658,9 @@ public class Pointer implements AutoCloseable {
             return true;
         } else if (obj == null) {
             return isNull();
-        } else if (obj.getClass() != getClass()) {
+        } else if (obj.getClass() != getClass()
+                && obj.getClass() != Pointer.class
+                && getClass() != Pointer.class) {
             return false;
         } else {
             Pointer other = (Pointer)obj;